feat: JPA 엔티티로 변환
