+++ src/main/java/gift/product/entity/Product.java
@@ -1,8 +1,18 @@
-package gift.product;
+package gift.product.entity;
 
+import static jakarta.persistence.GenerationType.IDENTITY;
+
+import gift.product.dto.ProductReqDto;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.Id;
+
+@Entity
 public class Product {
 
-    private final Long id;
+    @Id
+    @GeneratedValue(strategy = IDENTITY)
+    private  Long id;
     private String name;
     private Integer price;
     private String imageUrl;
@@ -12,6 +22,9 @@ public class Product {
         this.name = name;
         this.price = price;
         this.imageUrl = imageUrl;
+    }
+
+    protected Product() {
     }
 
     public Long getId() {
@@ -40,5 +53,11 @@ public class Product {
 
     public void setImageUrl(String imageUrl) {
         this.imageUrl = imageUrl;
+    }
+
+    public void update(ProductReqDto productReqDto) {
+        this.name = productReqDto.name();
+        this.price = productReqDto.price();
+        this.imageUrl = productReqDto.imageUrl();
     }
 }
