--- src/main/java/gift/wishlist/WishList.java
@@ -1,32 +1,0 @@
-package gift.wishlist;
-
-public class WishList {
-
-    private final Long id;
-    private Long memberId;
-    private Long productId;
-    private Integer quantity;
-
-    public WishList(Long id, Long memberId, Long productId, Integer quantity) {
-        this.id = id;
-        this.memberId = memberId;
-        this.productId = productId;
-        this.quantity = quantity;
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public Long getMemberId() {
-        return memberId;
-    }
-
-    public Long getProductId() {
-        return productId;
-    }
-
-    public Integer getQuantity() {
-        return quantity;
-    }
-}
